# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden:
    <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird entweder ein JSON-Array von JSON-Objekten, falls die Ressource eine Sammlung von Entitäten darstellt, oder ein JSON-Objekt, falls die Ressource eine einzelne Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
    </ul>
...
servers:
...
paths:
...
  /kuenstler:
    post:
      tags:
        - Anwender
      summary: Fügt einen Künstler hinzu.
      description: >-
        Ein Künstler wird mit den entsprechenden Werten hinzugefügt. Ein zugeordneter User wird ebenfalls angelegt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/user'
                - $ref: '#/components/schemas/kuenstler'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
...
  /bands:
    get:
      tags:
        - Anwender
      summary: Gibt Bands zurück.
      description: >-
        Alle Bands werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: name
          in: query
          description: 'Nur Bands, deren **Name** `name` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: genre
          in: query
          description: 'Nur Bands, denen das **Genre** `Genre` zugeordnet ist, werden zurückgegeben.'
          schema:
            type: string
        - name: gruendungsjahr
          in: query
          description: 'Nur Bands, dessen **Gründungsjahr** >= (größer-gleich) `gruendungsjahr` ist, werden zurückgegeben.'
          schema:
            type: integer
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        bandid:
                          type: integer
                          description: '**RowID der Band**.'
                      required:
                        - bandid
                    - $ref: '#/components/schemas/band'
    post:
      tags:
        - Authentifizierter und als Künstler authorisierter Anwender
      security:
        - basic authentication: []
      summary: Fügt eine Band hinzu.
      description: >-
        Eine Band wird mit den entsprechenden Werten hinzugefügt. **Der authentifizierte Künstler wird als Mitglied der Band eingetragen**.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/band'
                - type: object
                  properties:
                    genreid:
                      type: integer
                      description: '**RowID des Genres**.'
                  required:
                    - genreid
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
  /bands/{bandid}/kuenstler:
    post:
      tags:
        - Authentifizierter und als Künstler authorisierter Anwender
      security:
        - basic authentication: []
      summary: Fügt einer Band einen Künstler hinzu.
      description: >-
        Der Band mit **RowID** `bandid` wird ein Künstler hinzugefügt. **Der authentifizierte Künstler muss  Mitglied dieser Band sein**.
      parameters:
        - name: bandid
          in: path
          description: '**RowID der Band.**'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    kuenstlerid:
                      type: integer
                      description: '**RowID des Künstlers**.'
                  required:
                    - kuenstlerid
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
  /bands/{bandid}/genres:
    post:
      tags:
        - Authentifizierter und als Künstler authorisierter Anwender
      security:
        - basic authentication: []
      summary: Ordnet einer Band ein Genre zu.
      description: >-
        Der Band mit **RowID** `bandid` wird ein Genre zugeordnet. **Der authentifizierte Künstler muss Mitglied dieser Band sein**.
      parameters:
        - name: bandid
          in: path
          description: '**RowID der Band.**'
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - type: object
                  properties:
                    genreid:
                      type: integer
                      description: '**RowID des Genres.**.'
                  required:
                    - genreid
      responses:
        '204':
          $ref: '#/components/responses/204'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  securitySchemes:
    basic authentication:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource.
          schema:
            type: string
    '204':
      description: '**No Content**.'
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
    '404':
      description: '**Not Found**.'
  schemas:
    user:
      type: object
      properties:
        email:
          type: string
          description: '**E-Mail-Adresse** des Users.'
        passwort:
          type: string
          description: '**Passwort** des Users.'
        vorname:
          type: string
          description: '**Vorname** des Users.'
        nachname:
          type: string
          description: '**Nachname** des Users.'
      required:
        - email
        - vorname
        - nachname
        - passwort
    kuenstler:
      type: object
      properties:
        kuenstlername:
          type: string
          description: '**Künstlername** des Künstlers.'
...
    band:
      type: object
      properties:
        name:
          type: string
          description: '**Name** der Band.'
        gruendungsjahr:
          type: integer
          description: '**Gründungsjahr** der Band.'
      required:
        - name
        - gruendungsjahr
...